Feature/#10

+++ app/main.py
@@ -1,22 +1,24 @@
 from fastapi.middleware.cors import CORSMiddleware
 from contextlib import asynccontextmanager
 from datetime import timedelta
-from typing import AsyncGenerator, Optional
+from typing import  Optional
 import uuid
 from uuid import UUID
 import asyncpg
 from asyncpg import Connection
 from asyncpg.pool import Pool
-from fastapi import  Depends, FastAPI, File, Form, Query,UploadFile,HTTPException,Request
+from fastapi import  Depends, FastAPI, File, Form, Query,UploadFile,HTTPException
 import cloudinary,os
-from fastapi.security import OAuth2PasswordRequestForm, OAuth2PasswordBearer
+from fastapi.security import OAuth2PasswordRequestForm
 from auth import auth_user, create_access_token, get_current_user
-from database import DATABASE_URL
+from database import DATABASE_URL, get_db_conn
 from enums import ImageFormat
-from schemas import Image, Token
+from schemas import DBUser, Image, Token, User
 from cloudinary.uploader import upload,destroy
 from dotenv import load_dotenv
 from security import oauth2_scheme
+import hashlib
+import re
 load_dotenv()
 
 
@@ -42,6 +44,16 @@ async def lifespan(app: FastAPI):
             created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
         );
         """)
+
+        await conn.execute("""
+        CREATE TABLE IF NOT EXISTS users (
+            user_id UUID NOT NULL,
+            name VARCHAR NOT NULL,
+            password VARCHAR NOT NULL,
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            PRIMARY KEY (user_id)
+        );
+        """)
     yield
     # 後処理
     await app.state.db_pool.close()
@@ -62,12 +74,6 @@ cloudinary.config(
     api_key = str(os.getenv("CLOUDINARY_API_KEY")),
     api_secret = str(os.getenv("CLOUDINARY_API_SECRET"))
 )
-# ジェネレータ関数で共通化 依存性注入でconn取得部分を共通化
-async def get_db_conn(request: Request) -> AsyncGenerator[Connection, None]:
-    db_pool = request.app.state.db_pool
-    async with db_pool.acquire() as conn:
-        yield conn  # 非同期ジェネレータとして返す
-
 
 @app.get("/images")
 async def get_images(user_id: Optional[UUID] = Query(None),format: Optional[ImageFormat] = Query(None),conn:Connection = Depends(get_db_conn)): # Optionalが型でNone or Value Queryが入力時の話
@@ -98,10 +104,11 @@ async def get_image(public_id:UUID,conn:Connection = Depends(get_db_conn)):
 
 
 @app.post("/images")
-async def create_image(user_id:uuid.UUID=Form(...),title:str = Form(...),description:str = Form(...),image_file:UploadFile=File(...),conn:Connection=Depends(get_db_conn),auth_res = Depends(get_current_user)):
+async def create_image(title:str = Form(...),description:str = Form(...),image_file:UploadFile=File(...),conn:Connection=Depends(get_db_conn),current_user:DBUser = Depends(get_current_user)):
     # formリクエストを受けとって、cloudinaryにpubidをハッシュ化した画像をストア
     public_id = uuid.uuid4() # ストアする画像のUUID生成
     upload_contents = await image_file.read()
+    user_id = current_user.user_id
     try:
         # cloudinary操作
         
@@ -125,9 +132,19 @@ async def create_image(user_id:uuid.UUID=Form(...),title:str = Form(...),descrip
         raise HTTPException(status_code=500,detail=f"Database error: {e}")
     
 @app.delete("/images/{image_id}")
-async def delete_image(image_id:UUID,conn:Connection = Depends(get_db_conn),auth_res = Depends(get_current_user)):
+async def delete_image(image_id:UUID,conn:Connection = Depends(get_db_conn),current_user:DBUser = Depends(get_current_user)):
     
     # database 操作
+    row = await conn.fetchrow("SELECT * FROM images WHERE public_id = $1",image_id)
+    if row is None:
+        raise HTTPException(status_code=404,detail="Image not found")
+    
+    dict_row = dict(row)
+    if dict_row["user_id"] != current_user.user_id:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     
     res = await conn.execute("DELETE FROM images WHERE public_id = $1",image_id)
     if res == "DELETE 0":
@@ -140,16 +157,59 @@ async def delete_image(image_id:UUID,conn:Connection = Depends(get_db_conn),auth
     
     return {"detail":"Image deleted successfully"}
 
+@app.post("/users")
+async def create_user(name:str = Form(...),login_id:str=Form(...),password:str = Form(...),conn:Connection = Depends(get_db_conn)):
+    user_id = uuid.uuid4() # ユーザのUUIDを作成
+    hashed_password = hashlib.sha256(password.encode()).hexdigest()
+    if not re.fullmatch(r"^[a-zA-Z0-9_]+$",login_id): # login_id は半角英数字のみに限定
+        raise HTTPException(status_code=400,detail="login_id must contain only half-width letters, numbers, and underscores")
+    try:
+        res = await conn.execute("INSERT INTO users (user_id,name,login_id,password) VALUES ($1,$2,$3,$4)",user_id,name,login_id,hashed_password)
+    except asyncpg.UniqueViolationError:
+        raise HTTPException(status_code=409,detail="This login_id is already in use")
+
+    return {"user_id":user_id,"name":name,"login_id":login_id,"message":"User created successfully"}
+
+@app.get("/users")
+async def get_users(conn:Connection = Depends(get_db_conn)):
+    rows = await conn.fetch("SELECT * FROM users")
+    users = [] 
+    for row in rows: # password省く
+        user_dict = dict(row)
+        user_dict.pop("password",None)
+        users.append(user_dict)
+    
+    return users
+
+@app.get("/users/{user_uuid}")
+async def get_user(user_uuid:UUID,conn:Connection = Depends(get_db_conn)):
+    user_id = str(user_uuid)
+    row = await conn.fetchrow("SELECT * FROM users WHERE user_id = $1",user_id)
+    if row is None:
+        raise HTTPException(status_code=404,detail="User not found")
+    user_dict = dict(row)
+    user_dict.pop("password")
+    return user_dict
+
+@app.delete("/users/{user_uuid}")
+async def delete_user(user_uuid:UUID,conn:Connection = Depends(get_db_conn)):
+    user_id = str(user_uuid)
+    res = await conn.execute("DELETE FROM users WHERE user_id = $1",user_id)
+    command,count = res.split(" ")
+    if int(count) == 0:
+        raise HTTPException(status_code=404,detail="User not found")
+    
+    return {"message": "User deleted successfully"}
+
 @app.post("/login")
-async def login_for_access_token(form_data:OAuth2PasswordRequestForm = Depends())->Token:
+async def login_for_access_token(form_data:OAuth2PasswordRequestForm = Depends(),conn:Connection=Depends(get_db_conn)):
     # ログイン後トークンの作成
-    user_uuid = UUID(os.getenv("USER_ID"))
-    user = auth_user(user_id=user_uuid,user_name=form_data.username,password=form_data.password)
+    user= await auth_user(login_id=form_data.username,password=form_data.password,conn=conn)
     if not user:
         raise HTTPException(status_code=401,detail="Incorrect username or password",headers={"WWW-Authenticate": "Bearer"})
     minutes = float(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES") or 30)
     access_token_expires = timedelta(minutes=minutes)
     access_token = create_access_token(
-        data={"sub": user.user_name}, expires_delta=access_token_expires
+        data={"sub": user.login_id}, expires_delta=access_token_expires
     )
     return Token(access_token=access_token,token_type="bearer")
